// planning-idsdp-pps-frontend/src/app/main/modules/feasibility-study-management/models/cost-benefit-analysis.model.ts
export class CostBenefitAnalysisModel {
    financialAnalysis: string;
    financialNetPresentVal: number;
    financialBenefitCostRatio: number;
    financialInternalRateReturn: number;
    economicAnalysis: string;
    economicNetPresentVal: number;
    economicBenefitCostRatio: number;
    economicInternalRateReturn: number;
    attachment: File;
    fsrMasterId: number;

}